add red slide for regen. refactor some golem events
