+++ src/main/java/teamroots/emberroot/RegistryManager.java
@@ -1,6 +1,7 @@
 package teamroots.emberroot;
 import java.util.ArrayList;
 import java.util.List;
+import net.minecraft.client.Minecraft;
 import net.minecraft.entity.EnumCreatureType;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.biome.Biome;
@@ -17,6 +18,11 @@ import teamroots.emberroot.entity.deer.EntityDeer;
 import teamroots.emberroot.entity.deer.RenderDeer;
 import teamroots.emberroot.entity.fairy.EntityFairy;
 import teamroots.emberroot.entity.fairy.RenderFairy;
+import teamroots.emberroot.entity.golem.DamageGolem;
+import teamroots.emberroot.entity.golem.EntityAncientGolem;
+import teamroots.emberroot.entity.golem.EntityEmberProjectile;
+import teamroots.emberroot.entity.golem.RenderAncientGolem;
+import teamroots.emberroot.entity.golem.RenderEmberPacket;
 import teamroots.emberroot.entity.slime.EntityRainbowSlime;
 import teamroots.emberroot.entity.slime.RenderWaterSlime;
 import teamroots.emberroot.entity.sprout.EntitySprout;
@@ -27,37 +33,36 @@ public class RegistryManager {
     return (r * 65536 + g * 256 + b);
   }
   public static void registerAll() {
+    Roots.damage_ember = new DamageGolem();
     int id = 0;
     EntityRegistry.registerModEntity(new ResourceLocation(Const.MODID, "deer"), EntityDeer.class, "deer", id++, Roots.instance, 64, 1, true);
     EntityRegistry.registerEgg(new ResourceLocation(Const.MODID, "deer"), intColor(159, 132, 88), intColor(94, 77, 51));
     EntityRegistry.registerModEntity(new ResourceLocation(Const.MODID, "sprout"), EntitySprout.class, "sprout", id++, Roots.instance, 64, 1, true);
     EntityRegistry.registerEgg(new ResourceLocation(Const.MODID, "sprout"), intColor(136, 189, 33), intColor(165, 232, 42));
-    
-    
     EntityRegistry.registerModEntity(new ResourceLocation(Const.MODID, "fairy"), EntityFairy.class, "fairy", id++, Roots.instance, 64, 1, true);
     EntityRegistry.registerEgg(new ResourceLocation(Const.MODID, "fairy"), intColor(255, 208, 255), intColor(209, 255, 173));
-
     EntityRegistry.registerModEntity(new ResourceLocation(Const.MODID, "rainbowslime"), EntityRainbowSlime.class, "rainbowslime", id++, Roots.instance, 64, 1, true);
     EntityRegistry.registerEgg(new ResourceLocation(Const.MODID, "rainbowslime"), intColor(111, 208, 22), intColor(111, 255, 173));
-    
-    
+    EntityRegistry.registerModEntity(new ResourceLocation(Const.MODID, "ember_projectile"), EntityEmberProjectile.class, "ember_projectile", id++, Roots.instance, 64, 1, true);
+    EntityRegistry.registerModEntity(new ResourceLocation(Const.MODID, "ancient_golem"), EntityAncientGolem.class, "ancient_golem", id++, Roots.instance, 64, 1, true);
+    EntityRegistry.registerEgg(new ResourceLocation(Const.MODID, "ancient_golem"), intColor(48, 38, 35), intColor(79, 66, 61));
     List<Biome> biomesDeer = new ArrayList<Biome>();
     List<Biome> biomesFairy = new ArrayList<Biome>();
     List<Biome> biomesSprout = new ArrayList<Biome>();
+    List<Biome> biomesSlimes = new ArrayList<Biome>();
     for (BiomeEntry b : BiomeManager.getBiomes(BiomeType.COOL)) {
       biomesDeer.add(b.biome);
-      // biomesFairy.add(b.biome);
       biomesSprout.add(b.biome);
+      biomesSlimes.add(b.biome);
     }
     for (BiomeEntry b : BiomeManager.getBiomes(BiomeType.DESERT)) {
-      //biomesDeer.add(b.biome);
       biomesFairy.add(b.biome);
       biomesSprout.add(b.biome);
+      biomesSlimes.add(b.biome);
     }
     for (BiomeEntry b : BiomeManager.getBiomes(BiomeType.ICY)) {
       biomesDeer.add(b.biome);
-      // biomesFairy.add(b.biome);
-      //  biomesSprout.add(b.biome);
+      biomesSlimes.add(b.biome);
     }
     for (BiomeEntry b : BiomeManager.getBiomes(BiomeType.WARM)) {
       biomesDeer.add(b.biome);
@@ -67,9 +72,21 @@ public class RegistryManager {
     biomesFairy.addAll(BiomeManager.oceanBiomes);
     biomesDeer.addAll(BiomeManager.oceanBiomes);
     biomesSprout.addAll(BiomeManager.oceanBiomes);
-    EntityRegistry.addSpawn(EntityDeer.class, ConfigManager.deerSpawnWeight, 3, 7, EnumCreatureType.CREATURE, biomesDeer.toArray(new Biome[biomesDeer.size()]));
-    EntityRegistry.addSpawn(EntitySprout.class, ConfigManager.sproutSpawnWeight, 3, 7, EnumCreatureType.CREATURE, biomesSprout.toArray(new Biome[biomesSprout.size()]));
-    EntityRegistry.addSpawn(EntityFairy.class, ConfigManager.fairySpawnWeight, 3, 7, EnumCreatureType.CREATURE, biomesFairy.toArray(new Biome[biomesFairy.size()]));
+    EntityRegistry.addSpawn(EntityDeer.class, ConfigManager.deerSpawnWeight, 3, 7, EnumCreatureType.CREATURE, biomesDeer.toArray(new Biome[0]));
+    EntityRegistry.addSpawn(EntitySprout.class, ConfigManager.sproutSpawnWeight, 3, 7, EnumCreatureType.CREATURE, biomesSprout.toArray(new Biome[0]));
+    EntityRegistry.addSpawn(EntityFairy.class, ConfigManager.fairySpawnWeight, 3, 7, EnumCreatureType.CREATURE, biomesFairy.toArray(new Biome[0]));
+    EntityRegistry.addSpawn(EntityRainbowSlime.class, ConfigManager.slimeSpawnWeight, 1, 3, EnumCreatureType.MONSTER, biomesSlimes.toArray(new Biome[0]));
+    List<BiomeEntry> biomeGolems = new ArrayList<BiomeEntry>();
+    biomeGolems.addAll(BiomeManager.getBiomes(BiomeType.COOL));
+    biomeGolems.addAll(BiomeManager.getBiomes(BiomeType.DESERT));
+    biomeGolems.addAll(BiomeManager.getBiomes(BiomeType.ICY));
+    biomeGolems.addAll(BiomeManager.getBiomes(BiomeType.WARM));
+    List<Biome> biomes = new ArrayList<Biome>();
+    for (BiomeEntry b : biomeGolems) {
+      biomes.add(b.biome);
+    }
+    biomes.addAll(BiomeManager.oceanBiomes);
+    EntityRegistry.addSpawn(EntityAncientGolem.class, ConfigManager.golemSpawnWeight, 1, 1, EnumCreatureType.MONSTER, biomes.toArray(new Biome[0]));
   }
   @SideOnly(Side.CLIENT)
   @SubscribeEvent
@@ -78,5 +95,10 @@ public class RegistryManager {
     RenderingRegistry.registerEntityRenderingHandler(EntitySprout.class, new RenderSprout.Factory());
     RenderingRegistry.registerEntityRenderingHandler(EntityFairy.class, new RenderFairy.Factory());
     RenderingRegistry.registerEntityRenderingHandler(EntityRainbowSlime.class, new RenderWaterSlime.Factory());
+    RenderingRegistry.registerEntityRenderingHandler(EntityEmberProjectile.class, new RenderEmberPacket(Minecraft.getMinecraft().getRenderManager()));
+    RenderingRegistry.registerEntityRenderingHandler(EntityAncientGolem.class, new RenderAncientGolem.Factory());
   }
+  
+  
+  
 }
