+++ src/main/java/teamroots/emberroot/Roots.java
@@ -2,6 +2,7 @@ package teamroots.emberroot;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.init.Blocks;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.DamageSource;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
@@ -9,8 +10,7 @@ import net.minecraftforge.fml.common.Mod.Instance;
 import net.minecraftforge.fml.common.SidedProxy;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-import teamroots.emberroot.entity.golem.GolemRegistry;
+import net.minecraftforge.fml.relauncher.SideOnly; 
 import teamroots.emberroot.proxy.CommonProxy;
 
 @Mod(modid = Const.MODID, name = Roots.MODNAME, version = Roots.VERSION)
@@ -38,8 +38,11 @@ public class Roots {
     MinecraftForge.EVENT_BUS.register(new EventManager());
     MinecraftForge.EVENT_BUS.register(new ConfigManager());
     MinecraftForge.EVENT_BUS.register(new RegistryManager());
-    MinecraftForge.EVENT_BUS.register(new GolemRegistry());
     ConfigManager.init(event.getSuggestedConfigurationFile());
     proxy.preInit(event);
+  }
+  public static DamageSource damage_ember;
+  public static int intColor(int r, int g, int b) {
+    return (r * 65536 + g * 256 + b);
   }
 }
