+++ src/main/java/teamroots/emberroot/proxy/ClientProxy.java
@@ -1,8 +1,7 @@
 package teamroots.emberroot.proxy;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import teamroots.emberroot.entity.deer.ModelDeer;
-import teamroots.emberroot.entity.fairy.ModelFairy;
-import teamroots.emberroot.entity.golem.GolemRegistry;
+import teamroots.emberroot.entity.fairy.ModelFairy; 
 import teamroots.emberroot.entity.golem.ModelGolem;
 import teamroots.emberroot.entity.golem.ParticleRendererGolem;
 import teamroots.emberroot.entity.golem.RenderAncientGolem;
@@ -11,6 +10,7 @@ import teamroots.emberroot.entity.sprout.ModelSprout;
 import teamroots.emberroot.particle.ParticleRenderer;
 
 public class ClientProxy extends CommonProxy {
+  //TODO: merge these
   public static ParticleRenderer particleRenderer = new ParticleRenderer();
   public static ParticleRendererGolem particleRendererGolem = new ParticleRendererGolem();
   @Override
@@ -20,7 +20,6 @@ public class ClientProxy extends CommonProxy {
     ModelSprout.instance = new ModelSprout();
     ModelFairy.instance = new ModelFairy();
     ModelWaterSlime.instance = new ModelWaterSlime(16);
-    RenderAncientGolem.model = new ModelGolem();
-    GolemRegistry.registerEntityRendering();
+    RenderAncientGolem.model = new ModelGolem(); 
   }
 }
