+++ src/main/java/teamroots/emberroot/proxy/CommonProxy.java
@@ -4,8 +4,7 @@ import net.minecraftforge.fml.common.network.NetworkRegistry;
 import net.minecraftforge.fml.common.network.simpleimpl.SimpleNetworkWrapper;
 import net.minecraftforge.fml.relauncher.Side;
 import teamroots.emberroot.Const;
-import teamroots.emberroot.RegistryManager;
-import teamroots.emberroot.entity.golem.GolemRegistry;
+import teamroots.emberroot.RegistryManager; 
 import teamroots.emberroot.entity.golem.MessageEmberSizedBurstFX;
 
 public class CommonProxy {
@@ -14,6 +13,5 @@ public class CommonProxy {
     RegistryManager.registerAll();
     int id = 0;
     INSTANCE.registerMessage(MessageEmberSizedBurstFX.MessageHolder.class, MessageEmberSizedBurstFX.class, id++, Side.CLIENT);
-    GolemRegistry.registerAll();
   }
 }
